requirements.txt
# Core dependencies
torch>=2.1.0
transformers>=4.36.0
tokenizers>=0.15.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ONNX Runtime (GPU)
onnxruntime-gpu>=1.16.0
onnx>=1.15.0

# API Server
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6

# Training
datasets>=2.16.0
accelerate>=0.25.0
scikit-learn>=1.3.0

# HTTP Client (for distillation)
httpx>=0.26.0

# Utilities
numpy>=1.24.0
tqdm>=4.66.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Development
black>=23.12.0
ruff>=0.1.0
